refactor: refactoring Int to Long for all Id

See merge request s11-final/S11P31B206!84

+++ backend/homeggu-goods/src/main/java/com/homeggu/domain/goods/GoodsRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface GoodsRepository extends JpaRepository <Goods, Integer> {
-    List<Goods> findBySalesBoardId(Integer salesBoardId);
+public interface GoodsRepository extends JpaRepository <Goods, Long> {
+    List<Goods> findBySalesBoardId(Long salesBoardId);
 }
